update parser and some fixes
