+++ packages/eslint-plugin-svelte/package.json
@@ -66,7 +66,7 @@
     "postcss-safe-parser": "^6.0.0",
     "postcss-selector-parser": "^7.0.0",
     "semver": "^7.6.3",
-    "svelte-eslint-parser": "^1.0.0-next.1"
+    "svelte-eslint-parser": "^1.0.0-next.3"
   },
   "devDependencies": {
     "@babel/core": "^7.26.0",
