+++ packages/eslint-plugin-svelte/src/type-defs/estree.d.ts
@@ -1,0 +1,85 @@
+// IMPORTANT!
+// This file has been automatically generated,
+// in order to update its content execute "pnpm run update"
+//
+// Replace type information to use "@typescript-eslint/types" instead of "estree".
+//
+declare module 'estree' {
+	import type { TSESTree } from '@typescript-eslint/types';
+
+	export type Node = TSESTree.Node;
+	export type Program = TSESTree.Program;
+	export type Expression = TSESTree.Expression;
+	export type Statement = TSESTree.Statement;
+	export type Pattern = TSESTree.DestructuringPattern;
+	export type AccessorProperty = TSESTree.AccessorProperty;
+	export type ArrayExpression = TSESTree.ArrayExpression;
+	export type ArrayPattern = TSESTree.ArrayPattern;
+	export type ArrowFunctionExpression = TSESTree.ArrowFunctionExpression;
+	export type AssignmentExpression = TSESTree.AssignmentExpression;
+	export type AssignmentPattern = TSESTree.AssignmentPattern;
+	export type AwaitExpression = TSESTree.AwaitExpression;
+	export type BinaryExpression = TSESTree.BinaryExpression;
+	export type BlockStatement = TSESTree.BlockStatement;
+	export type BreakStatement = TSESTree.BreakStatement;
+	export type CallExpression = TSESTree.CallExpression;
+	export type CatchClause = TSESTree.CatchClause;
+	export type ChainExpression = TSESTree.ChainExpression;
+	export type ClassBody = TSESTree.ClassBody;
+	export type ClassDeclaration = TSESTree.ClassDeclaration;
+	export type ClassExpression = TSESTree.ClassExpression;
+	export type ConditionalExpression = TSESTree.ConditionalExpression;
+	export type ContinueStatement = TSESTree.ContinueStatement;
+	export type DebuggerStatement = TSESTree.DebuggerStatement;
+	export type DoWhileStatement = TSESTree.DoWhileStatement;
+	export type EmptyStatement = TSESTree.EmptyStatement;
+	export type ExportAllDeclaration = TSESTree.ExportAllDeclaration;
+	export type ExportDefaultDeclaration = TSESTree.ExportDefaultDeclaration;
+	export type ExportNamedDeclaration = TSESTree.ExportNamedDeclaration;
+	export type ExportSpecifier = TSESTree.ExportSpecifier;
+	export type ExpressionStatement = TSESTree.ExpressionStatement;
+	export type ForInStatement = TSESTree.ForInStatement;
+	export type ForOfStatement = TSESTree.ForOfStatement;
+	export type ForStatement = TSESTree.ForStatement;
+	export type FunctionDeclaration = TSESTree.FunctionDeclaration;
+	export type FunctionExpression = TSESTree.FunctionExpression;
+	export type Identifier = TSESTree.Identifier;
+	export type IfStatement = TSESTree.IfStatement;
+	export type ImportDeclaration = TSESTree.ImportDeclaration;
+	export type ImportDefaultSpecifier = TSESTree.ImportDefaultSpecifier;
+	export type ImportExpression = TSESTree.ImportExpression;
+	export type ImportNamespaceSpecifier = TSESTree.ImportNamespaceSpecifier;
+	export type ImportSpecifier = TSESTree.ImportSpecifier;
+	export type LabeledStatement = TSESTree.LabeledStatement;
+	export type Literal = TSESTree.Literal;
+	export type LogicalExpression = TSESTree.LogicalExpression;
+	export type MemberExpression = TSESTree.MemberExpression;
+	export type MetaProperty = TSESTree.MetaProperty;
+	export type MethodDefinition = TSESTree.MethodDefinition;
+	export type NewExpression = TSESTree.NewExpression;
+	export type ObjectExpression = TSESTree.ObjectExpression;
+	export type ObjectPattern = TSESTree.ObjectPattern;
+	export type PrivateIdentifier = TSESTree.PrivateIdentifier;
+	export type Property = TSESTree.Property;
+	export type PropertyDefinition = TSESTree.PropertyDefinition;
+	export type RestElement = TSESTree.RestElement;
+	export type ReturnStatement = TSESTree.ReturnStatement;
+	export type SequenceExpression = TSESTree.SequenceExpression;
+	export type SpreadElement = TSESTree.SpreadElement;
+	export type Super = TSESTree.Super;
+	export type SwitchCase = TSESTree.SwitchCase;
+	export type SwitchStatement = TSESTree.SwitchStatement;
+	export type TaggedTemplateExpression = TSESTree.TaggedTemplateExpression;
+	export type TemplateElement = TSESTree.TemplateElement;
+	export type TemplateLiteral = TSESTree.TemplateLiteral;
+	export type ThisExpression = TSESTree.ThisExpression;
+	export type ThrowStatement = TSESTree.ThrowStatement;
+	export type TryStatement = TSESTree.TryStatement;
+	export type UnaryExpression = TSESTree.UnaryExpression;
+	export type UpdateExpression = TSESTree.UpdateExpression;
+	export type VariableDeclaration = TSESTree.VariableDeclaration;
+	export type VariableDeclarator = TSESTree.VariableDeclarator;
+	export type WhileStatement = TSESTree.WhileStatement;
+	export type WithStatement = TSESTree.WithStatement;
+	export type YieldExpression = TSESTree.YieldExpression;
+}
