+++ packages/eslint-plugin-svelte/tools/update-types-for-node.ts
@@ -9,7 +9,7 @@ const __dirname = path.dirname(new URL(import.meta.url).pathname);
 // const filename = fileURLToPath(import.meta.url)
 const dirname = __dirname; // path.dirname(filename)
 const typesForNodeFilename = path.join(dirname, '../src/types-for-node.ts');
-const estreeFilename = path.join(dirname, '../typings/estree/index.d.ts');
+const estreeFilename = path.join(dirname, '../src/type-defs/estree.d.ts');
 const { visitorKeys } = parseForESLint('');
 
 const esNextNodeNames = ['Decorator', 'ImportAttribute', 'StaticBlock'];
@@ -32,7 +32,7 @@ const estreeCode = [
 //
 // Replace type information to use "@typescript-eslint/types" instead of "estree".
 //
-
+declare module 'estree' {
 import type { TSESTree } from "@typescript-eslint/types"
 
 export type Node = TSESTree.Node
@@ -112,5 +112,6 @@ for (const nodeType of svelteNodeNames.filter((k) => !esSvelteNodeNames.includes
 }
 typesForNodeCode.push(`}`);
 
+estreeCode.push(`}`);
 void writeAndFormat(typesForNodeFilename, typesForNodeCode.join('\n'));
 void writeAndFormat(estreeFilename, estreeCode.join('\n'));
